refactor(backend): move graphql server to separate file
