+++ Graphql_Server/src/index.ts
@@ -1,9 +1,4 @@
-import { ApolloServer } from "@apollo/server";
 import { expressMiddleware } from "@apollo/server/express4";
-import { ApolloServerPluginDrainHttpServer } from "@apollo/server/plugin/drainHttpServer";
-import { makeExecutableSchema } from "@graphql-tools/schema";
-import { WebSocketServer } from "ws";
-import { useServer } from "graphql-ws/lib/use/ws";
 
 import cors from "cors";
 import express from "express";
@@ -11,67 +6,17 @@ import { createServer } from "http";
 
 import { getUser } from "./lib/utils.js";
 import { setupDatabase } from "./lib/db.js";
-import resolvers from "./resolvers/index.js";
-import typeDefs from "./typeDefs/index.js";
 import { GraphQLError } from "graphql";
+import createApolloServer from "./lib/apolloServer.js";
 
 const PORT = 4000;
-
-const schema = makeExecutableSchema({ typeDefs, resolvers });
 
 const app = express();
 const httpServer = createServer(app);
 
-const wsServer = new WebSocketServer({
-    server: httpServer,
-    path: "/",
-});
-
-const serverCleanup = useServer(
-    {
-        schema,
-        context: async (ctx) => {
-            const user = await getUser(ctx);
-
-            if (!user)
-                throw new GraphQLError("User is not authenticated", {
-                    extensions: {
-                        code: "UNAUTHENTICATED",
-                        http: { status: 401 },
-                    },
-                });
-
-            return { user };
-        },
-        onConnect: async (ctx) => {
-            const user = await getUser(ctx);
-            if (!user) return false;
-            console.log("Connected WebSocket");
-        },
-        onDisconnect(ctx, code, reason) {
-            console.log("Disconnected WebSocket");
-        },
-    },
-    wsServer
-);
-
-const server = new ApolloServer({
-    schema,
-    plugins: [
-        ApolloServerPluginDrainHttpServer({ httpServer }),
-        {
-            async serverWillStart() {
-                return {
-                    async drainServer() {
-                        await serverCleanup.dispose();
-                    },
-                };
-            },
-        },
-    ],
-});
-
+const server = createApolloServer(httpServer);
 await server.start();
+
 app.use(
     "/",
     cors(),
